About to remove more glitch glue
