+++ Example/PaymentSheet Example/PaymentSheet Example/ExampleWalletButtonsView.swift
@@ -253,9 +253,9 @@ class ExampleWalletButtonsModel: ObservableObject {
         )
 
         return PaymentSheet.ShopPayConfiguration(shippingAddressRequired: true,
-                                                 lineItems: [.init(name: "product 1", amount: 1099)],
-                                                 shippingRates: [.init(id: "id1", amount: 1099, displayName: "Overnight", deliveryEstimate: .init(minimum: singleBusinessDay, maximum: singleBusinessDay)),
-                                                                 .init(id: "id2", amount: 0, displayName: "Free", deliveryEstimate: .init(minimum: fiveBusinessDays, maximum: sevenBusinessDays)),
+                                                 lineItems: [.init(name: "Golden Potato", amount: 500), .init(name: "Silver Potato", amount: 345), .init(name: "Tax", amount: 200)],
+                                                 shippingRates: [.init(id: "1day", amount: 1099, displayName: "Overnight", deliveryEstimate: .init(minimum: singleBusinessDay, maximum: singleBusinessDay)),
+                                                                 .init(id: "free", amount: 0, displayName: "Free", deliveryEstimate: .init(minimum: fiveBusinessDays, maximum: sevenBusinessDays)),
                                                                 ],
                                                  shopId: "shop_1234",
                                                  handlers: handlers)
