+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/ECE/ECEIndexHtml.swift
@@ -3,7 +3,6 @@
 //  StripePaymentSheet
 //
 
-
 let ECEHTML = """
 <!DOCTYPE html>
 <html>
@@ -16,10 +15,10 @@ let ECEHTML = """
     <div id="express-checkout-element">
       <!-- Express Checkout Element will be inserted here -->
     </div>
-    
+
     <script>
-    
-    
+
+
     function getStripePublishableKey() {
       return "pk_test_51RUTiSAs6uch2mqQune4yYMgnaPTI8z7AuCS9CPb5zaDQuUsje3qsRZKwgjDND3DTwvKVz6aSWYFy36FVA7iyn7h00QbaV5A9S";
     }
@@ -87,9 +86,9 @@ let ECEHTML = """
     function initializeStripeElements() {
 
       console.log("Initializing stripe elements with options", ELEMENTS_OPTIONS);
-      
+
       let stripe, elements, expressCheckoutElement;
-      
+
       try {
         stripe = window.Stripe(getStripePublishableKey(), {});
         console.log("Created Stripe");
@@ -98,7 +97,7 @@ let ECEHTML = """
         logECEEvent(`❌ Failed to create Stripe instance: ${error.message}`);
         throw error;
       }
-      
+
       try {
         elements = stripe.elements(ELEMENTS_OPTIONS);
         console.log("Created Elements");
@@ -108,7 +107,7 @@ let ECEHTML = """
         logECEEvent(`Elements options were: ${JSON.stringify(ELEMENTS_OPTIONS)}`);
         throw error;
       }
-      
+
       try {
         expressCheckoutElement = elements.create(
           "expressCheckout",
@@ -143,10 +142,10 @@ let ECEHTML = """
 
       expressCheckoutElement.on("click", async function (event) {
         logECEEvent(`Click received with event:\n${hashToString(event)}`);
-        
+
         try {
           logECEEvent("Using Native ECE Click Handler...");
-          
+
           // Extract only serializable data from the event
           // The event object contains functions (resolve, complete) that can't be cloned
           // for message passing, so we only send the data we need
@@ -154,30 +153,30 @@ let ECEHTML = """
             walletType: event.walletType,
             expressPaymentType: event.expressPaymentType
           };
-          
+
           // Call native API to handle the click
           // The native handler will provide all configuration like:
           // - billingAddressRequired, emailRequired, phoneNumberRequired
           // - shippingAddressRequired, allowedShippingCountries
           // - business name, line items, etc.
           const resolvePayload = await window.NativeECE.handleClick(eventData);
-          
+
           logECEEvent(`Native ECE Response:\n${hashToString(resolvePayload)}`);
-          
+
           // Resolve the event with the payload from native
           event.resolve(resolvePayload);
-          
+
           // Handle overlay removal if requested by native
           if (resolvePayload.disableOverlay) {
             // Remove the overlay applied by ECE
             document.querySelectorAll('div[style*="z-index: 9999999"]')[0]?.remove();
             document.body.style.overflow = "auto";
           }
-          
+
         } catch (error) {
           logECEEvent(`Error handling ECE click: ${error.message}`);
           console.error("ECE click error:", error);
-          
+
           // On error, throw to let Stripe handle it
           throw error;
         }
@@ -191,7 +190,7 @@ let ECEHTML = """
         );
         shippingCountry = event.address.country;
         shippingState = event.address.state;
-        
+
         // Transform address to match native API format
         const shippingAddress = {
           address1: event.address.addressLine ? event.address.addressLine[0] : "",
@@ -206,53 +205,53 @@ let ECEHTML = """
           postalCode: event.address.postalCode || "",
           provinceCode: event.address.state || ""
         };
-        
+
         try {
           // Check if native API is available
           if (window.NativeShipping && window.NativeShipping.calculateShipping) {
             logECEEvent("Using Native Shipping API...");
-            
+
             // Call native API
             const response = await window.NativeShipping.calculateShipping(shippingAddress);
-            
+
             logECEEvent(`Native API Response:\n${hashToString(response)}`);
-            
+
             // Check if merchant rejected the address
             if (response.merchantDecision === "rejected") {
               logECEEvent(`Shipping rejected: ${response.error || "Address not serviceable"}`);
               event.reject();
               return;
             }
-            
+
             // Use the response from native API
             SHIPPING_RATES = response.shippingRates || [];
-            
+
             // Update line items if provided
             if (response.lineItems) {
               LINE_ITEMS = response.lineItems;
             }
-            
+
             // Update the total amount from the response
             if (response.totalAmount) {
               AMOUNT_TOTAL = response.totalAmount;
               elements.update({ amount: AMOUNT_TOTAL });
             }
-            
+
             event.resolve({
               shippingRates: SHIPPING_RATES,
               lineItems: LINE_ITEMS
             });
-            
+
           } else {
             // Resolve with an error
             logECEEvent("Native Shipping API not available");
             event.reject();
           }
-          
+
         } catch (error) {
           logECEEvent(`Error calculating shipping: ${error.message}`);
           console.error("Shipping calculation error:", error);
-          
+
           // On error, reject the shipping address
           event.reject();
         }
@@ -263,42 +262,42 @@ let ECEHTML = """
         console.log("Shipping rate change event:", event);
         console.log("Current AMOUNT_TOTAL:", AMOUNT_TOTAL);
         console.log("Current SHIPPING_RATES:", SHIPPING_RATES);
-        
+
         try {
           // Check if native API is available
           console.log("Checking for native rate API...");
           console.log("window.NativeShipping:", window.NativeShipping);
           console.log("calculateShippingRateChange function:", window.NativeShipping?.calculateShippingRateChange);
-          
+
           if (window.NativeShipping && window.NativeShipping.calculateShippingRateChange) {
             logECEEvent("Using Native Shipping Rate API...");
-            
+
             // Call native API to validate and calculate new amount
             const response = await window.NativeShipping.calculateShippingRateChange(event.shippingRate, AMOUNT_TOTAL);
-            
+
             logECEEvent(`Native Rate API Response:\n${hashToString(response)}`);
-            
+
             // Check if merchant rejected the rate
             if (response.merchantDecision === "rejected") {
               logECEEvent(`Shipping rate rejected: ${response.error || "Invalid shipping rate"}`);
               event.reject();
               return;
             }
-            
+
             // Update the amount with the value from native
             if (response.updatedAmount) {
               console.log(`updating amount to ${response.updatedAmount}`);
               elements.update({ amount: response.updatedAmount });
             }
-            
+
             event.resolve();
-            
+
           }
-          
+
         } catch (error) {
           logECEEvent(`Error validating shipping rate: ${error.message}`);
           console.error("Shipping rate validation error:", error);
-          
+
           // On error, reject the shipping rate
           event.reject();
         }
@@ -318,7 +317,7 @@ let ECEHTML = """
           )}\nShippingRate:\n${hashToString(event.shippingRate)}}}`
         );
         console.log("confirm payment event triggered", event);
-        
+
         const { error: submitError } = await elements.submit();
         if (submitError) {
           console.log("submit error");
@@ -331,7 +330,7 @@ let ECEHTML = """
           // Check if native API is available
           if (window.NativePayment && window.NativePayment.confirmPayment) {
             logECEEvent("Using Native Payment API...");
-            
+
             // Prepare payment details for native processing
             const paymentDetails = {
               billingDetails: event.billingDetails,
@@ -341,15 +340,15 @@ let ECEHTML = """
               captureMethod: captureMethod,
               //paymentMethod: paymentMethod,
             };
-            
+
             // Call native API to create and confirm payment
             const response = await window.NativePayment.confirmPayment(paymentDetails);
-            
+
             logECEEvent(`Native Payment API Response:\n${hashToString(response)}`);
-            
+
             // Use the response from native API
             const { clientSecret, paymentIntentId } = response;
-            
+
             if (mode === "payment") {
               const confirmedPI = await stripe.confirmPayment({
                 elements,
@@ -374,9 +373,9 @@ let ECEHTML = """
             } else {
               console.log(`Invalid mode onConfirm ${mode}.`);
             }
-            
+
           }
-          
+
         } catch (error) {
           logECEEvent(`Error confirming payment: ${error.message}`);
           console.error("Payment confirmation error:", error);
@@ -389,7 +388,7 @@ let ECEHTML = """
       // Also log to browser console
       console.log("[ECE Event]", msg);
     }
-    
+
     // Function to initialize everything - called from Swift or automatically
     function initializeApp() {
       getItems()
