+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/ECE/ECEViewController.swift
@@ -8,11 +8,10 @@
 //  This provides native async/await support without manual request ID management
 //
 
-import UIKit
-import WebKit
 @_spi(STP) import StripeCore
 @_spi(STP) import StripeUICore
-
+import UIKit
+import WebKit
 
 // Protocol for handling Express Checkout Element events
 protocol ExpressCheckoutWebviewDelegate: AnyObject {
@@ -29,7 +28,7 @@ enum ExpressCheckoutError: LocalizedError {
     case shippingServiceFailure(underlying: Error)
     case missingRequiredField(field: String)
     case paymentConfirmationFailed(reason: String)
-    
+
     var errorDescription: String? {
         switch self {
         case .invalidShippingAddress(let details):
@@ -48,247 +47,141 @@ enum ExpressCheckoutError: LocalizedError {
 
 @available(iOS 16.0, *)
 class ECEViewController: UIViewController {
-    
+
     private var webView: WKWebView!
     private var popupWebView: WKWebView?
     private var popupWindow: UIWindow?
-    
+
     // Store fetched items data
     private var lineItems: [[String: Any]] = []
     private var amountTotal: Int = 0
-    
+
     // Store default shipping rates
-    private var defaultShippingRates: [[String: Any]] = []
-    
+    private var shippingRates: [[String: Any]] = []
+
     // Delegate for Express Checkout Element events
     weak var expressCheckoutWebviewDelegate: ExpressCheckoutWebviewDelegate?
-    
+
     // We normally hide Express Checkout Element in WKWebViews (as many don't handle popups correctly)
     // Fake Safari's UA to disable that behavior
     static let FakeSafariUserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 18_1 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/18.1.1 Mobile/15E148 Safari/604.1"
-    
+
     override func loadView() {
         // Create main view
         view = UIView()
         view.backgroundColor = .systemBackground
-        
+
     }
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         setupNavigationBar()
-        
+
         // Add a loading spinner
         let spinner = UIActivityIndicatorView(style: .large)
         spinner.color = .systemGray
         spinner.translatesAutoresizingMaskIntoConstraints = false
         spinner.startAnimating()
-        
+
         view.addSubview(spinner)
         NSLayoutConstraint.activate([
             spinner.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            spinner.centerYAnchor.constraint(equalTo: view.centerYAnchor)
+            spinner.centerYAnchor.constraint(equalTo: view.centerYAnchor),
         ])
-        
+
         // Fetch items and setup webview asynchronously
         Task {
             await loadDataAndSetupWebView()
         }
     }
-    
+
     // New async method to load data and setup web view
     private func loadDataAndSetupWebView() async {
         // Fetch items first
         await fetchItems()
-        
+
         // Then fetch default shipping rates
         await fetchDefaultShippingRates()
-        
+
         // Setup web view on main thread
         await MainActor.run {
             setupWebView()
-            
+
             // Add the tiny 1x1 webview as a hidden subview
             view.addSubview(webView)
             loadECE()
         }
     }
-    
+
     private func setupWebView() {
         let configuration = WKWebViewConfiguration()
-        
+
         // Enable JavaScript
         configuration.defaultWebpagePreferences.allowsContentJavaScript = true
         configuration.preferences.javaScriptCanOpenWindowsAutomatically = true
-        
+
         // Setup message handlers for the bridge
         let contentController = WKUserContentController()
-        
+
         // Add message handlers for different types of messages
         // Use regular handlers for one-way messages
         contentController.add(self, name: "ready")
         contentController.add(self, name: "error")
         contentController.add(self, name: "consoleLog")
-        
+
         // Use reply handlers for request/response messages (iOS 14+)
         contentController.addScriptMessageHandler(self, contentWorld: .page, name: "calculateShipping")
         contentController.addScriptMessageHandler(self, contentWorld: .page, name: "calculateShippingRateChange")
         contentController.addScriptMessageHandler(self, contentWorld: .page, name: "confirmPayment")
         contentController.addScriptMessageHandler(self, contentWorld: .page, name: "handleECEClick")
-        
+
         configuration.userContentController = contentController
-        
+
         // Inject JavaScript to capture messages and set up the bridge
         let bridgeScript = """
-
-
-        // Function to simulate click on shop-pay-payment-request-button in iframes
-        function findAndClickShopPayButton() {
-            function searchInFrame(frameWindow, depth = 0) {
-                if (depth > 5) return false; // Prevent infinite recursion
-                
-                try {
-                    const button = frameWindow.document.querySelector('shop-pay-payment-request-button');
-                    if (button) {
-                        console.log('Found shop-pay-payment-request-button at depth', depth);
-                        console.log('Attempting to click Shop Pay button - this may trigger a popup window');
-                        
-                        // Try multiple click simulation methods
-                        button.click();
-                        
-                        // Also dispatch synthetic click events
-                        const clickEvent = new MouseEvent('click', {
-                            bubbles: true,
-                            cancelable: true,
-                            view: frameWindow
-                        });
-                        button.dispatchEvent(clickEvent);
-                        
-                        // Try triggering any onclick handlers
-                        if (button.onclick) {
-                            button.onclick();
-                        }
-                        
-                        // Also try focus and enter key for accessibility
-                        button.focus();
-                        const enterEvent = new KeyboardEvent('keydown', {
-                            key: 'Enter',
-                            code: 'Enter',
-                            keyCode: 13,
-                            bubbles: true
-                        });
-                        button.dispatchEvent(enterEvent);
-                        
-                        window.webkit.messageHandlers.stripeMessage.postMessage({
-                            type: 'shopPayButtonClicked',
-                            depth: depth,
-                            timestamp: Date.now(),
-                            buttonFound: true,
-                            origin: frameWindow.location.origin,
-                            message: 'Shop Pay button clicked - watch for popup window'
-                        });
-                        
-                        return true;
-                    }
-                    
-                    // Search in child iframes
-                    const iframes = frameWindow.document.querySelectorAll('iframe');
-                    for (let iframe of iframes) {
-                        try {
-                            if (searchInFrame(iframe.contentWindow, depth + 1)) {
-                                return true;
-                            }
-                        } catch(e) {
-                            // Cross-origin iframe, skip
-                            console.log('Cross-origin iframe at depth', depth + 1);
-                        }
-                    }
-                } catch(e) {
-                    console.log('Error searching frame at depth', depth, ':', e.message);
-                }
-                
-                return false;
-            }
-            
-            const found = searchInFrame(window);
-            
-            if (!found) {
-        // Log
-            }
-        }
-        
-        // Wait 1 second after page load, then try to find and click the shop pay button
-        setTimeout(() => {
-            console.log('Searching for shop-pay-payment-request-button...');
-            //findAndClickShopPayButton();
-        }, 1000);
-        
-        
         // Inject line items from native
         window.NATIVE_LINE_ITEMS = \(lineItemsAsJSON());
         window.NATIVE_AMOUNT_TOTAL = \(amountTotal);
-        
+
         window.NativeShipping = {
             // Direct async/await API to calculate shipping
             calculateShipping: async function(shippingAddress) {
-                try {
-                    // iOS 14+ automatically handles the reply as a promise
-                    return await window.webkit.messageHandlers.calculateShipping.postMessage({
-                        shippingAddress: shippingAddress,
-                        timestamp: Date.now()
-                    });
-                } catch(e) {
-                    throw new Error('Failed to communicate with native: ' + e.message);
-                }
+                return await window.webkit.messageHandlers.calculateShipping.postMessage({
+                    shippingAddress: shippingAddress
+                });
             },
-            
+
             // Direct async/await API to validate shipping rate change
             calculateShippingRateChange: async function(shippingRate, currentAmount) {
-                try {
-                    return await window.webkit.messageHandlers.calculateShippingRateChange.postMessage({
-                        shippingRate: shippingRate,
-                        currentAmount: currentAmount,
-                        timestamp: Date.now()
-                    });
-                } catch(e) {
-                    throw new Error('Failed to communicate with native: ' + e.message);
-                }
+                return await window.webkit.messageHandlers.calculateShippingRateChange.postMessage({
+                    shippingRate: shippingRate,
+                    currentAmount: currentAmount
+                });
             }
         };
-        
+
         window.NativePayment = {
             // Direct async/await API to confirm payment
             confirmPayment: async function(paymentDetails) {
-                try {
-                    return await window.webkit.messageHandlers.confirmPayment.postMessage({
-                        paymentDetails: paymentDetails,
-                        timestamp: Date.now()
-                    });
-                } catch(e) {
-                    throw new Error('Failed to communicate with native: ' + e.message);
-                }
+                return await window.webkit.messageHandlers.confirmPayment.postMessage({
+                    paymentDetails: paymentDetails
+                });
             }
         };
-        
+
         window.NativeECE = {
             // Direct async/await API to handle ECE click events
             handleClick: async function(eventData) {
-                try {
-                    return await window.webkit.messageHandlers.handleECEClick.postMessage({
-                        eventData: eventData,
-                        timestamp: Date.now()
-                    });
-                } catch(e) {
-                    throw new Error('Failed to communicate with native: ' + e.message);
-                }
+                return await window.webkit.messageHandlers.handleECEClick.postMessage({
+                    eventData: eventData
+                });
             }
         };
-        
+
         // Notify that the bridge is ready
         try {
             window.webkit.messageHandlers.ready.postMessage({
                 type: 'bridgeReady',
-                timestamp: Date.now(),
                 userAgent: navigator.userAgent,
                 url: window.location.href,
                 origin: window.location.origin,
@@ -299,7 +192,7 @@ class ECEViewController: UIViewController {
             // Ignore errors if webkit handlers not available
         }
         """
-        
+
         // Add console interceptor as early as possible
         let consoleInterceptor = """
         // Early console interceptor to catch all logs
@@ -311,7 +204,7 @@ class ECEViewController: UIViewController {
                 info: console.info,
                 debug: console.debug
             };
-            
+
             function formatArgs(args) {
                 return Array.from(args).map(arg => {
                     if (typeof arg === 'object') {
@@ -324,7 +217,7 @@ class ECEViewController: UIViewController {
                     return String(arg);
                 }).join(' ');
             }
-            
+
             ['log', 'error', 'warn', 'info', 'debug'].forEach(method => {
                 console[method] = function(...args) {
                     originalConsole[method].apply(console, args);
@@ -332,7 +225,6 @@ class ECEViewController: UIViewController {
                         window.webkit.messageHandlers.consoleLog.postMessage({
                             level: method,
                             message: formatArgs(args),
-                            timestamp: Date.now(),
                             stackTrace: method === 'error' ? (new Error()).stack : undefined
                         });
                     } catch(e) {}
@@ -340,13 +232,13 @@ class ECEViewController: UIViewController {
             });
         })();
         """
-        
+
         let earlyScript = WKUserScript(source: consoleInterceptor, injectionTime: .atDocumentStart, forMainFrameOnly: false)
         contentController.addUserScript(earlyScript)
-        
+
         let script = WKUserScript(source: bridgeScript, injectionTime: .atDocumentEnd, forMainFrameOnly: false)
         contentController.addUserScript(script)
-        
+
         // Create a tiny 1x1 pixel webview positioned at origin
         // It's 500x500 for now for debugging
         webView = WKWebView(frame: CGRect(x: 0, y: 0, width: 500, height: 500), configuration: configuration)
@@ -359,40 +251,38 @@ class ECEViewController: UIViewController {
         #endif
         webView.isHidden = false // Keep it technically visible but tiny
         webView.alpha = 1.00 // Make it nearly transparent // don't do this for now
-        
+
         webView.customUserAgent = Self.FakeSafariUserAgent
     }
-    
+
     private func setupNavigationBar() {
-        title = "Stripe Checkout Bridge"
-        
-        // Add refresh button
+        title = "Checkout"
+
+        // Add refresh button, back, and forward for debugging. Let's get rid of these before ship
         navigationItem.rightBarButtonItem = UIBarButtonItem(
             barButtonSystemItem: .refresh,
             target: self,
             action: #selector(refreshWebView)
         )
-        
-        // Add back/forward buttons if needed
         navigationItem.leftBarButtonItems = [
             UIBarButtonItem(title: "Back", style: .plain, target: self, action: #selector(goBack)),
-            UIBarButtonItem(title: "Forward", style: .plain, target: self, action: #selector(goForward))
+            UIBarButtonItem(title: "Forward", style: .plain, target: self, action: #selector(goForward)),
         ]
     }
-    
+
     private func loadECE() {
         webView.loadHTMLString(ECEHTML, baseURL: URL(string: "https://pay.stripe.com")!)
     }
-    
+
     // Fetch items from the Glitch API - now async
     private func fetchItems() async {
         let url = URL(string: "https://unexpected-dune-list.glitch.me/items")!
-        
+
         print("🛒 Fetching items from: \(url)")
-        
+
         do {
             let (data, _) = try await URLSession.shared.data(from: url)
-            
+
             guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
                   let items = json["items"] as? [[String: Any]],
                   let total = json["total"] as? Int else {
@@ -400,82 +290,64 @@ class ECEViewController: UIViewController {
                 // Use default values on error
                 self.lineItems = [
                     ["name": "Golden Potato", "amount": 500],
-                    ["name": "Silver Potato", "amount": 345]
+                    ["name": "Silver Potato", "amount": 345],
                 ]
                 self.amountTotal = 1045
                 return
             }
-            
+
             print("✅ Fetched \(items.count) items with total: \(total)")
-            
+
             // Store the fetched items
             self.lineItems = items.map { item in
                 return [
                     "name": item["name"] ?? "",
-                    "amount": item["amount"] ?? 0
+                    "amount": item["amount"] ?? 0,
                 ]
             }
             self.amountTotal = total
-            
+
         } catch {
             print("❌ Failed to fetch items: \(error)")
             // Use default values on error
             self.lineItems = [
                 ["name": "Golden Potato", "amount": 500],
-                ["name": "Silver Potato", "amount": 345]
+                ["name": "Silver Potato", "amount": 345],
             ]
             self.amountTotal = 1045
         }
     }
-    
-    // Fetch default shipping rates for CA, US - now async
+
     private func fetchDefaultShippingRates() async {
         // Create a default shipping address for California, US
         let defaultAddress = ShippingAddress(provinceCode: "CA", countryCode: "US")
-        
+
         print("📦 Fetching default shipping rates for CA, US...")
-        
+
         do {
             let rates = try await ShippingService.shared.fetchShippingRates(for: defaultAddress)
             print("✅ Fetched \(rates.count) default shipping rates")
-            
+
             // Convert ShippingRate objects to dictionaries for storage
-            self.defaultShippingRates = rates.map { rate in
+            self.shippingRates = rates.map { rate in
                 return [
                     "id": rate.id,
                     "displayName": rate.displayName,
                     "amount": rate.amount,
-                    "deliveryEstimate": rate.deliveryEstimate
+                    "deliveryEstimate": rate.deliveryEstimate,
                 ]
             }
-            
+
             // Log the fetched rates
             for rate in rates {
                 print("   📋 \(rate.displayName): $\(Double(rate.amount)/100.0) - \(rate.deliveryEstimate)")
             }
-            
+
         } catch {
             print("⚠️ Failed to fetch default shipping rates: \(error)")
-            print("   Using fallback shipping rates")
-            
-            // Use fallback rates if API fails
-            self.defaultShippingRates = [
-                [
-                    "id": "standard",
-                    "displayName": "Standard Shipping",
-                    "amount": 0,
-                    "deliveryEstimate": "5-7 Business Days"
-                ],
-                [
-                    "id": "express", 
-                    "displayName": "Express Shipping",
-                    "amount": 800,
-                    "deliveryEstimate": "2-3 Business Days"
-                ]
-            ]
         }
     }
-    
+
     // Helper method to convert line items to JSON string
     private func lineItemsAsJSON() -> String {
         guard let data = try? JSONSerialization.data(withJSONObject: lineItems, options: []),
@@ -484,17 +356,17 @@ class ECEViewController: UIViewController {
         }
         return jsonString
     }
-    
+
     @objc private func refreshWebView() {
         webView.reload()
     }
-    
+
     @objc private func goBack() {
         if webView.canGoBack {
             webView.goBack()
         }
     }
-    
+
     @objc private func goForward() {
         if webView.canGoForward {
             webView.goForward()
@@ -508,12 +380,12 @@ extension ECEViewController: WKScriptMessageHandler {
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
         let messageBody = message.body
         let timestamp = DateFormatter.logFormatter.string(from: Date())
-        
+
         switch message.name {
         case "ready":
             print("✅ [\(timestamp)] Bridge Ready:")
             // TODO: Some timer here to make sure we get ready in time or log an error
-            
+
         case "error":
             print("❌ [\(timestamp)] Bridge Error:")
             if let errorDict = messageBody as? [String: Any] {
@@ -521,14 +393,14 @@ extension ECEViewController: WKScriptMessageHandler {
             } else {
                 print("   Error: \(messageBody)")
             }
-            
+
         case "consoleLog":
             if let logDict = messageBody as? [String: Any],
                let level = logDict["level"] as? String,
                let logMessage = logDict["message"] as? String {
                 let emoji = logEmojiForLevel(level)
                 print("\(emoji) [\(timestamp)] JS Console.\(level): \(logMessage)")
-                
+
                 // Optionally print stack trace for errors
                 if level == "error", let stackTrace = logDict["stackTrace"] as? String {
                     let lines = stackTrace.split(separator: "\n").prefix(5) // Show first 5 lines of stack
@@ -537,12 +409,12 @@ extension ECEViewController: WKScriptMessageHandler {
                     }
                 }
             }
-            
+
         default:
             print("🔍 [\(timestamp)] Unknown message type '\(message.name)': \(messageBody)")
         }
     }
-    
+
     private func printMessageDetails(_ messageDict: [String: Any]) {
         for (key, value) in messageDict.sorted(by: { $0.key < $1.key }) {
             if let nestedDict = value as? [String: Any] {
@@ -568,7 +440,7 @@ extension ECEViewController: WKScriptMessageHandler {
             }
         }
     }
-    
+
     private func logEmojiForLevel(_ level: String) -> String {
         switch level.lowercased() {
         case "error": return "❌"
@@ -582,8 +454,8 @@ extension ECEViewController: WKScriptMessageHandler {
 }
 
 struct BridgeError: Error {
-    var localizedDescription: String
-    
+    public var localizedDescription: String
+
     init(_ localizedDescription: String) {
         self.localizedDescription = localizedDescription
     }
@@ -591,7 +463,7 @@ struct BridgeError: Error {
 // MARK: - WKScriptMessageHandlerWithReply (iOS 14+)
 @available(iOS 16.0, *)
 extension ECEViewController: WKScriptMessageHandlerWithReply {
-    
+
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage, replyHandler: @escaping (Any?, String?) -> Void) {
         Task {
             do {
@@ -602,7 +474,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
             }
         }
     }
-    
+
     func handleMessage(message: WKScriptMessage) async throws -> Any? {
         let messageBody = message.body
         let timestamp = DateFormatter.logFormatter.string(from: Date())
@@ -612,7 +484,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
         guard let messageDict = messageBody as? [String: Any] else {
             throw BridgeError("Invalid message format \(messageBody)")
         }
-        
+
         switch message.name {
         case "calculateShipping":
             print("🚚 [\(timestamp)] Calculate Shipping Request:")
@@ -621,7 +493,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
             } else {
                 throw BridgeError("Invalid calculateShipping message format")
             }
-            
+
         case "calculateShippingRateChange":
             print("📦 [\(timestamp)] Calculate Shipping Rate Change Request:")
             if let shippingRate = messageDict["shippingRate"] as? [String: Any] {
@@ -629,7 +501,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
             } else {
                 throw BridgeError("Invalid calculateShippingRateChange message format")
             }
-            
+
         case "confirmPayment":
             print("💳 [\(timestamp)] Confirm Payment Request (iOS 14+ Reply):")
             if let paymentDetails = messageDict["paymentDetails"] as? [String: Any] {
@@ -637,7 +509,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
             } else {
                 throw BridgeError("Invalid confirmPayment message format")
             }
-            
+
         case "handleECEClick":
             print("👆 [\(timestamp)] ECE Click Event:")
             if let eventData = messageDict["eventData"] as? [String: Any] {
@@ -645,7 +517,7 @@ extension ECEViewController: WKScriptMessageHandlerWithReply {
             } else {
                 throw BridgeError("Invalid handleECEClick message format")
             }
-            
+
         default:
             throw BridgeError("Unknown message type: \(message.name)")
         }
@@ -658,14 +530,14 @@ extension ECEViewController: WKNavigationDelegate {
     func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
         print("🚀 Navigation started: \(webView.url?.absoluteString ?? "unknown")")
     }
-    
+
     func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
         print("✅ Navigation finished: \(webView.url?.absoluteString ?? "unknown")")
-        
+
         // If this is the main page load, initialize the app with our native data
         if webView.url?.absoluteString.contains("pay.stripe.com") == true {
             // Call the JavaScript initializeApp() function now that native data is injected
-            webView.evaluateJavaScript("initializeApp()") { result, error in
+            webView.evaluateJavaScript("initializeApp()") { _, error in
                 if let error = error {
                     print("❌ Failed to call initializeApp(): \(error)")
                     // Bail with error
@@ -675,11 +547,11 @@ extension ECEViewController: WKNavigationDelegate {
             }
         }
     }
-    
+
     func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
         print("❌ Navigation failed: \(error.localizedDescription)")
     }
-    
+
     func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
         decisionHandler(.allow)
     }
@@ -689,9 +561,9 @@ extension ECEViewController: WKNavigationDelegate {
 @available(iOS 16.0, *)
 extension ECEViewController: WKUIDelegate {
     func webView(_ webView: WKWebView, createWebViewWith configuration: WKWebViewConfiguration, for navigationAction: WKNavigationAction, windowFeatures: WKWindowFeatures) -> WKWebView? {
-        
+
         print("🪟 Popup requested for URL: \(navigationAction.request.url?.absoluteString ?? "unknown")")
-        
+
         // Use the provided configuration directly (this fixes the configuration error)
         // Create popup with full screen bounds (unlike the 1x1 main webview)
         popupWebView = WKWebView(frame: UIScreen.main.bounds, configuration: configuration)
@@ -702,57 +574,57 @@ extension ECEViewController: WKUIDelegate {
             popupWebView?.isInspectable = true
         }
         #endif
-        
+
         // Set the same Safari user agent for popups
         popupWebView?.customUserAgent = Self.FakeSafariUserAgent
-        
+
         // Create popup window
         if let windowScene = view.window?.windowScene {
             popupWindow = UIWindow(windowScene: windowScene)
         }
-        
+
         let popupViewController = UIViewController()
         popupViewController.view = popupWebView
-        popupViewController.title = "Shop Pay"
-        
+        popupViewController.title = "Checkout"
+
         let navController = UINavigationController(rootViewController: popupViewController)
-        
+
         // Add close button to popup
         popupViewController.navigationItem.leftBarButtonItem = UIBarButtonItem(
             barButtonSystemItem: .done,
             target: self,
             action: #selector(closePopup)
         )
-        
+
         // Add refresh button to popup
         popupViewController.navigationItem.rightBarButtonItem = UIBarButtonItem(
             barButtonSystemItem: .refresh,
             target: self,
             action: #selector(refreshPopup)
         )
-        
+
         popupWindow?.rootViewController = navController
         popupWindow?.makeKeyAndVisible()
-        
+
         return popupWebView
     }
-    
+
     @objc private func closePopup() {
         popupWindow?.isHidden = true
         popupWindow = nil
         popupWebView = nil
     }
-    
+
     @objc private func refreshPopup() {
         popupWebView?.reload()
     }
-    
+
     func webViewDidClose(_ webView: WKWebView) {
         if webView == popupWebView {
             closePopup()
         }
     }
-    
+
     // Handle generic JavaScript alerts/confirms/prompts
     func webView(_ webView: WKWebView, runJavaScriptAlertPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping () -> Void) {
         let alert = UIAlertController(title: frame.request.url?.host() ?? "", message: message, preferredStyle: .alert)
@@ -761,7 +633,7 @@ extension ECEViewController: WKUIDelegate {
         })
         present(alert, animated: true)
     }
-    
+
     func webView(_ webView: WKWebView, runJavaScriptConfirmPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (Bool) -> Void) {
         let alert = UIAlertController(title: frame.request.url?.host() ?? "", message: message, preferredStyle: .alert)
         alert.addAction(UIAlertAction(title: String.Localized.ok, style: .default) { _ in
@@ -772,7 +644,7 @@ extension ECEViewController: WKUIDelegate {
         })
         present(alert, animated: true)
     }
-    
+
     func webView(_ webView: WKWebView, runJavaScriptTextInputPanelWithPrompt prompt: String, defaultText: String?, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (String?) -> Void) {
         let alert = UIAlertController(title: frame.request.url?.host() ?? "", message: prompt, preferredStyle: .alert)
         alert.addTextField { textField in
@@ -800,7 +672,7 @@ extension DateFormatter {
 // MARK: - Express Checkout Delegate Implementation
 @available(iOS 16.0, *)
 extension ECEViewController: ExpressCheckoutWebviewDelegate {
-    
+
     // Implementation using the new ShippingService
     func webView(_ webView: WKWebView, didReceiveShippingAddressChange shippingAddress: [String: Any]) async throws -> [String: Any] {
         print("🏠 Received shipping address change:")
@@ -808,32 +680,32 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         print("   City: \(shippingAddress["city"] ?? "")")
         print("   State: \(shippingAddress["provinceCode"] ?? "")")
         print("   ZIP: \(shippingAddress["postalCode"] ?? "")")
-        
+
         // Convert dictionary to ShippingAddress struct
         guard let address = ShippingAddress(from: shippingAddress) else {
             print("❌ Failed to parse shipping address")
             throw ExpressCheckoutError.invalidShippingAddress(details: "Failed to parse address from provided data")
         }
-        
+
         // Use ShippingService to calculate shipping
         do {
             let shippingResponse = try await ShippingService.shared.calculateShipping(for: address)
-            
+
             // If the response is rejected, throw an error instead
             if shippingResponse.merchantDecision == "rejected" {
                 let errorMessage = shippingResponse.error ?? "Shipping not available"
                 throw ExpressCheckoutError.invalidShippingAddress(details: errorMessage)
             }
-            
+
             // Create a modified response using our fetched line items
             var modifiedResponse = shippingResponse.asDictionary
-            
+
             // Use our fetched line items instead of the hardcoded ones from ShippingService
             if shippingResponse.merchantDecision == "accepted" && !self.lineItems.isEmpty {
                 modifiedResponse["lineItems"] = self.lineItems
                 modifiedResponse["totalAmount"] = self.amountTotal
             }
-            
+
             // Convert ShippingResponse to dictionary and send back
             return modifiedResponse
         } catch {
@@ -845,45 +717,43 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
             throw ExpressCheckoutError.shippingServiceFailure(underlying: error)
         }
     }
-    
+
     // Implementation for shipping rate changes
     func webView(_ webView: WKWebView, didReceiveShippingRateChange shippingRate: [String: Any]) async throws -> [String: Any] {
         print("📦 Received shipping rate change:")
         print("   ID: \(shippingRate["id"] ?? "")")
         print("   Display Name: \(shippingRate["displayName"] ?? "")")
         print("   Amount: \(shippingRate["amount"] ?? 0)")
-        
+
         // Extract shipping rate info
         let shippingId = shippingRate["id"] as? String ?? ""
         let shippingAmount = shippingRate["amount"] as? Int ?? 0
-        
+
         // You can validate the shipping rate against your known rates here
         // Use the default shipping rates we fetched at startup plus common fallback IDs
-        let defaultRateIds = defaultShippingRates.compactMap { $0["id"] as? String }
-        let fallbackRateIds = ["free", "3day", "2day", "standard", "express"]
-        let knownRateIds = defaultRateIds + fallbackRateIds
-        
-        if knownRateIds.contains(shippingId) {
+        let rateIds = shippingRates.compactMap { $0["id"] as? String }
+
+        if rateIds.contains(shippingId) {
             // Calculate the new total amount using our fetched base amount
             let baseAmount = amountTotal // Use the fetched amount
             let updatedAmount = baseAmount + shippingAmount
-            
+
             print("✅ Shipping rate accepted. New total: \(updatedAmount)")
-            
+
             let response: [String: Any] = [
                 "merchantDecision": "accepted",
                 "updatedAmount": updatedAmount,
-                "shippingRateAccepted": true
+                "shippingRateAccepted": true,
             ]
             return response
         } else {
             // Reject unknown shipping rates
             print("❌ Unknown shipping rate ID: \(shippingId)")
-            
+
             throw ExpressCheckoutError.invalidShippingRate(rateId: shippingId)
         }
     }
-    
+
     // Implementation for ECE click events
     func webView(_ webView: WKWebView, didReceiveECEClick event: [String: Any]) async throws -> [String: Any] {
         print("👆 Received ECE click event:")
@@ -893,15 +763,15 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         if let expressPaymentType = event["expressPaymentType"] as? String {
             print("   Express Payment Type: \(expressPaymentType)")
         }
-        
+
         // Native configuration - these values would typically come from your app's settings
         // or be determined based on business logic
-        
+
         // Example: You could read these from UserDefaults, app configuration, or business rules
         // let billingAddressRequired = UserDefaults.standard.bool(forKey: "requireBillingAddress")
         // let shippingAddressRequired = order.requiresShipping
         // let allowedShippingCountries = BusinessLogic.getShippingCountries()
-        
+
         let billingAddressRequired = true
         let emailRequired = true
         let phoneNumberRequired = true
@@ -909,7 +779,7 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         let businessName = "" // Optional - could be fetched from app settings
         let allowedShippingCountries = ["US", "CA"] // Could be based on your business regions
         let disableOverlay = false // Could be a user preference
-        
+
         print("   Native Configuration:")
         print("     billingAddressRequired: \(billingAddressRequired)")
         print("     emailRequired: \(emailRequired)")
@@ -918,8 +788,8 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         print("     businessName: \(businessName)")
         print("     allowedShippingCountries: \(allowedShippingCountries)")
         print("     disableOverlay: \(disableOverlay)")
-        print("     defaultShippingRates: \(defaultShippingRates.count) rates loaded")
-        
+        print("     ShippingRates: \(shippingRates.count) rates loaded")
+
         // Build the resolve payload
         var resolvePayload: [String: Any] = [
             "lineItems": lineItems,
@@ -928,36 +798,36 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
             "phoneNumberRequired": phoneNumberRequired,
             "shippingAddressRequired": shippingAddressRequired,
             "allowedShippingCountries": allowedShippingCountries,
-            "disableOverlay": disableOverlay
+            "disableOverlay": disableOverlay,
         ]
-        
+
         // Add shipping rates if shipping is required
         if shippingAddressRequired {
             // Use default shipping rates that were fetched at startup
             resolvePayload["shippingRates"] = defaultShippingRates
         }
-        
+
         // Add business info if provided
         if !businessName.isEmpty {
             resolvePayload["business"] = [
                 "name": businessName
             ]
         }
-        
+
         print("✅ Resolving ECE click with payload:")
         printMessageDetails(resolvePayload)
-        
+
         return resolvePayload
     }
-    
+
     func webView(_ webView: WKWebView, didReceiveECEConfirmation paymentDetails: [String: Any]) async throws -> [String: Any] {
         print("💳 Received ECE confirmation request")
-        
+
         // Extract payment details
         guard let billingDetails = paymentDetails["billingDetails"] as? [String: Any] else {
             throw ExpressCheckoutError.missingRequiredField(field: "billingDetails")
         }
-        
+
 //        let shippingAddress = paymentDetails["shippingAddress"] as? [String: Any]
         let shippingRate = paymentDetails["shippingRate"] as? [String: Any]
         let selectedShippingId = shippingRate?["id"] as? String
@@ -965,7 +835,7 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         let captureMethod = paymentDetails["captureMethod"] as? String ?? "automatic"
 //        let paymentMethod = paymentDetails["paymentMethod"] as? [String: Any]
 //        let createPaymentMethodEnabled = paymentDetails["createPaymentMethodEnabled"] as? Bool ?? false
-        
+
         print("   Mode: \(mode)")
         print("   Capture Method: \(captureMethod)")
         if let email = billingDetails["email"] as? String {
@@ -974,36 +844,36 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
         if let selectedShippingId = selectedShippingId {
             print("   Selected Shipping: \(selectedShippingId)")
         }
-        
+
         // Create payment intent
         let paymentData = try await createPaymentIntent(
             mode: mode,
             captureMethod: captureMethod,
             selectedShippingId: selectedShippingId
         )
-        
+
         guard let clientSecret = paymentData["secret"] as? String,
               let paymentIntentId = paymentData["paymentIntentId"] as? String else {
             throw ExpressCheckoutError.paymentConfirmationFailed(reason: "Invalid payment intent response - missing secret or ID")
         }
-        
+
         print("✅ PaymentIntent created: \(paymentIntentId)")
-        
+
         // Build the response
         let response: [String: Any] = [
             "clientSecret": clientSecret,
             "paymentIntentId": paymentIntentId,
             "mode": mode,
             "requiresAction": false,
-            "status": "requires_confirmation"
+            "status": "requires_confirmation",
         ]
-        
+
         // Get payment intent details after a short delay
         Task {
             try? await Task.sleep(nanoseconds: 2_000_000_000) // 2 seconds
             await getPaymentIntentDetails(paymentIntentId: paymentIntentId)
         }
-        
+
         return response
     }
 }
@@ -1011,11 +881,11 @@ extension ECEViewController: ExpressCheckoutWebviewDelegate {
 // MARK: - Payment Methods
 @available(iOS 16.0, *)
 extension ECEViewController {
-    
+
     // Handle payment confirmation with reply handler (iOS 14+)
     private func handlePaymentConfirmationWithReply(paymentDetails: [String: Any], replyHandler: @escaping (Any?, String?) -> Void) async {
         print("💳 Processing payment confirmation...")
-        
+
         // If delegate is set, let it handle the ECE confirmation
         if let delegate = expressCheckoutWebviewDelegate {
             do {
@@ -1030,7 +900,7 @@ extension ECEViewController {
             await handleDirectPaymentConfirmation(paymentDetails: paymentDetails, replyHandler: replyHandler)
         }
     }
-    
+
     // Direct payment confirmation without delegate
     @available(iOS 16.0, *)
     private func handleDirectPaymentConfirmation(paymentDetails: [String: Any], replyHandler: @escaping (Any?, String?) -> Void) async {
@@ -1051,38 +921,38 @@ extension ECEViewController {
                 captureMethod: captureMethod,
                 selectedShippingId: selectedShippingId
             )
-            
+
             guard let clientSecret = paymentData["secret"] as? String,
                   let paymentIntentId = paymentData["paymentIntentId"] as? String else {
                 replyHandler(nil, "Invalid payment intent response")
                 return
             }
-            
+
             print("✅ PaymentIntent created: \(paymentIntentId)")
-            
+
             // Return the response directly via reply handler
             let response: [String: Any] = [
                 "clientSecret": clientSecret,
                 "paymentIntentId": paymentIntentId,
                 "mode": mode,
                 "requiresAction": false,
-                "status": "requires_confirmation"
+                "status": "requires_confirmation",
             ]
-            
+
             replyHandler(response, nil)
-            
+
             // Get payment intent details after a short delay
             Task {
                 try? await Task.sleep(nanoseconds: 2_000_000_000) // 2 seconds
                 await getPaymentIntentDetails(paymentIntentId: paymentIntentId)
             }
-            
+
         } catch {
             print("❌ Failed to create payment intent: \(error)")
             replyHandler(nil, error.localizedDescription)
         }
     }
-    
+
     // Create payment intent by calling the Glitch endpoint - now async
     private func createPaymentIntent(
         mode: String,
@@ -1093,42 +963,41 @@ extension ECEViewController {
         var request = URLRequest(url: url)
         request.httpMethod = "POST"
         request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        
+
         let body: [String: Any] = [
             "mode": mode,
             "captureMethod": captureMethod,
-            "selectedShippingId": selectedShippingId ?? NSNull()
+            "selectedShippingId": selectedShippingId ?? NSNull(),
         ]
-        
+
         request.httpBody = try JSONSerialization.data(withJSONObject: body)
-        
+
         let (data, _) = try await URLSession.shared.data(for: request)
-        
+
         guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
             throw ExpressCheckoutError.paymentConfirmationFailed(reason: "Invalid server response")
         }
-        
+
         return json
     }
-    
+
     // Get payment intent details - now async
     private func getPaymentIntentDetails(paymentIntentId: String) async {
         let url = URL(string: "https://unexpected-dune-list.glitch.me/intent/\(paymentIntentId)")!
-        
+
         do {
             let (data, _) = try await URLSession.shared.data(from: url)
-            
+
             guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
                 print("❌ Invalid payment intent response")
                 return
             }
-            
+
             print("📋 PaymentIntent details:")
             printMessageDetails(json)
         } catch {
             print("❌ Failed to get payment intent details: \(error)")
         }
     }
-    
-}
 
+}
