+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetConfiguration.swift
@@ -591,6 +591,10 @@ extension PaymentSheet {
         /// Whether or not to collect the customer's shipping address
         public let shippingAddressRequired: Bool
 
+        /// By default, the Express Checkout Element allows all countries for shipping.
+        /// You can specify which countries are allowed for shipping in the Express Checkout Element with a list of two-letter country codes
+        public let allowedShippingCountries: [String]
+
         /// An array of LineItem objects. These LineItems are shown as line items in the payment interface, if line items are supported. You can represent discounts as negative amount LineItems.
         public let lineItems: [LineItem]
 
@@ -610,6 +614,7 @@ extension PaymentSheet {
             lineItems: [LineItem],
             shippingRates: [ShippingRate],
             shopId: String,
+            allowedShippingCountries: [String] = [],
             handlers: Handlers? = nil
         ) {
             self.billingAddressRequired = billingAddressRequired
@@ -618,6 +623,7 @@ extension PaymentSheet {
             self.lineItems = lineItems
             self.shippingRates = shippingRates
             self.shopId = shopId
+            self.allowedShippingCountries = allowedShippingCountries
             self.handlers = handlers
         }
     }
